requirements.txt
Appium-Python-Client==3.1.0
pytest==7.4.3